+++ src/core/layers/tests/unit/js/layer2dRasterWms.spec.js
@@ -197,6 +197,31 @@ describe("src/core/js/layers/layer2dRasterWms.js", () => {
                 useFetchForWMS: false
             });
         });
+
+        it("should return the layer params, overwriteWmsLoadfunction is true", () => {
+            store.getters = {
+                isModuleAvailable: () => false
+            };
+            Config.overwriteWmsLoadfunction = true;
+
+            const wmsLayer = new Layer2dRasterWms(localAttributes);
+
+            expect(wmsLayer.getLayerParams(localAttributes)).to.deep.equals({
+                format: "image/png",
+                gfiAsNewWindow: false,
+                gfiAttributes: "showAll",
+                gfiTheme: "default",
+                infoFormat: "text/xml",
+                layers: "test_layers",
+                name: "test_name",
+                opacity: 0.9,
+                typ: "wms",
+                zIndex: 1,
+                featureCount: 5,
+                gfiThemeSettings: undefined,
+                useFetchForWMS: true
+            });
+        });
     });
 
     describe("getOptions", () => {