+++ src/modules/login/components/LoginComponent.vue
@@ -10,6 +10,9 @@ export default {
         ...mapGetters(["isMobile"]),
         ...mapGetters("Modules/Login", ["loggedIn", "screenName", "email", "iconLogin", "iconLogged"])
     },
+    created () {
+        Config.overwriteWmsLoadfunction = true;
+    },
     async mounted () {
         if (!await this.isLoggedIn()) {
             this.openLoginWindow();