+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Issue #1524: draw_old: Added documentation for drawSquareSettings.
 - CopyrightConstraints: Added Parameter `useLayerCswUrl` to use the CSW interface specified in the layers.
 - LayerTree: Added LayerComponentIconCustom to provide a custom action button for layers that can execute Masterportal actions.
+- WMS: Added parameter `overwriteWmsLoadfunction` used in config.js to enable to overwrite the global wms load function.
 - WMS-Time:
     - The attributes `dimensionRange` and `dimensionRegex` can now be used to filter the values ​​that should be displayed in the time slider.
     - The attibute `staticDimensions` can now be configured and sent to the service when the WMS time layer is loaded.